Rename utils.sealed_date module (and tests) to utils.partial_date
